+++ tests/integration/client/dlm_migration_client.py
@@ -2,10 +2,11 @@
 
 import requests
 
+from ska_dlm.typer_types import JsonObjectOption
 from tests.integration.client.exception_handler import dlm_raise_for_status
 
 MIGRATION_URL = ""
-TOKEN: str = None
+TOKEN: str | None = None
 
 
 def copy_data_item(
@@ -16,6 +17,7 @@ def copy_data_item(
     destination_name: str = "",
     destination_id: str = "",
     path: str = "",
+    metadata: JsonObjectOption = None,
 ) -> dict:
     """Copy a data_item from source to destination.
 
@@ -40,6 +42,8 @@ def copy_data_item(
         the destination storage, by default ""
     path
         the destination path relative to storage root, by default ""
+    metadata
+        Metadata associated with the migration. Can be Null.
 
     Returns
     -------